chore: moving more code
