+++ packages/icheck-ts/src/cli.ts
@@ -1,24 +1,17 @@
 import process from "node:process";
-import { TinyCliCommand, arg, tinyCliMain } from "@hiogawa/tiny-cli";
+import { TinyCliCommand, tinyCliMain } from "@hiogawa/tiny-cli";
 import { version as packageVersion } from "../package.json";
-import { type ExportUsage, run } from "./runner";
+import { runCommand, runnerArgs } from "./runner";
 
 const command = new TinyCliCommand(
   {
     program: "icheck-ts",
     version: packageVersion,
     description: "Lint import and export usages",
-    args: {
-      files: arg.stringArray("Typescript files to lint"),
-      cache: arg.boolean("enable caching"),
-      ignore: arg.string("RegExp pattern to ignore export names", {
-        optional: true,
-      }),
-      debug: arg.boolean("Debug output"),
-    },
+    args: runnerArgs,
   },
-  ({ args }) => {
-    const result = run(args.files, { cache: args.cache });
+  async ({ args }) => {
+    const { result, unused } = await runCommand(args);
 
     if (args.debug) {
       console.log(new Map(result.exportUsages));
@@ -31,21 +24,6 @@ const command = new TinyCliCommand(
       }
       process.exitCode = 1;
     }
-
-    const ignoreRegExp = args.ignore && new RegExp(args.ignore);
-    const ignoreComment = "icheck-ignore";
-
-    function isUsedExport(e: ExportUsage): boolean {
-      return Boolean(
-        e.used ||
-          (ignoreRegExp && e.name.match(ignoreRegExp)) ||
-          e.comment.includes(ignoreComment)
-      );
-    }
-
-    const unused = [...result.exportUsages]
-      .map(([k, vs]) => [k, vs.filter((v) => !isUsedExport(v))] as const)
-      .filter(([_k, vs]) => vs.length > 0);
 
     if (unused.length > 0) {
       console.log("* Unused exports");
